requirements.txt
discord.py[voice]==2.5.2
yt-dlp==2025.8.11
python-dotenv==1.0.1
flask==3.0.3
psycopg2-binary==2.9.11
pynacl==1.5.0
aiohttp==3.10.5
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.40.6
trafilatura==1.12.2
psutil==6.0.0
pytz==2024.1
